Conflicts:
	nbproject/build-impl.xml
	nbproject/genfiles.properties
	nbproject/project.properties
	nbproject/project.xml

+++ src/java/admins/Borrarusuario_cofaa.java
@@ -11,6 +11,7 @@ package admins;
  */
 
 import com.opensymphony.xwork2.ActionSupport;
+import java.sql.ResultSet;
 
 public class Borrarusuario_cofaa extends ActionSupport{
     
@@ -28,13 +29,25 @@ public class Borrarusuario_cofaa extends ActionSupport{
     }
     @Override
     public String execute() throws Exception {
+        int a = 0;
         LoginBean lb = new LoginBean();
         lb.getConnection();
-        int val2=lb.executeUpdate("DELETE FROM evaluador WHERE id_usuario='" + getId_usuario() + "'");
-        int val=lb.executeUpdate("DELETE FROM usuario WHERE id_usuario='"+getId_usuario()+"'");
-        lb.closeConnection();
-        if (val>0) return "borrar_exitoso";
-        else return "borrar_fallo"; 
+        ResultSet rs = lb.executeQuery("SELECT COUNT(*) FROM evaluador");
+        rs.next();
+        a = rs.getInt(1);
+        if(a == 1)
+            return "Al_menos_uno";
+        else{
+            int val2=lb.executeUpdate("DELETE FROM evaluador WHERE id_usuario='" + getId_usuario() + "'");
+            int val=lb.executeUpdate("DELETE FROM usuario WHERE id_usuario='"+getId_usuario()+"'");
+            lb.closeConnection();
+            if (val>0) 
+                return "borrar_exitoso";
+            else 
+                return "borrar_fallo"; 
+        }
+            
+        
     }   
 }
